Added 'read leaderbord' REST API - Modified Game's code to send REST request for leaderboard. - Installed Rest Framework in Django

+++ coregame/models.py
@@ -11,4 +11,4 @@ class GameScoring(models.Model):
         on_delete=models.CASCADE,
     )
     timing = models.PositiveIntegerField(null=True)
-    board_size = models.PositiveSmallIntegerField(null=True)
+    board_size = models.PositiveIntegerField(null=True)
